pyproject.toml
[project]
name = "multiagentcoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=7.1.1",
    "langchain-community>=0.3.13",
    "langchain>=0.3.13",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.60",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
    "wolframalpha>=5.1.3",
    "langchain-experimental>=0.3.4",
    "arxiv>=2.1.3",
    "wikipedia>=1.4.0",
    "stackapi>=0.3.1",
    "docker>=7.1.0",
    "rich>=13.9.4",
    "rizaio>=0.4.0",
    "pubchempy>=1.0.4",
    "requests>=2.32.3",
    "rdkit>=2024.9.1",
    "chem-bench",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "sympy>=1.13.1",
    "deepchem>=2.8.0",
    "networkx>=3.4.2",
    "pydot>=3.0.3",
    "xmltodict>=0.14.2",
    "pymupdf>=1.25.1",
]

[tool.uv.sources]
chem-bench = { git = "https://github.com/lamalab-org/chem-bench.git" }
